LayoutTests/http/wpt/cross-origin-options/cross-origin-options-header.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Basic testing for Cross-Origin-Options HTTP header</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
</head>
<body>
<script>

// Test frame has a subframe so we expect an indexed property with index 0.
crossOriginPropertyNames.push('0');

function checkIframePropertyValues(w)
{
    assert_equals(w.parent, window, "'parent' property value");
    assert_equals(w.top, window, "'top' property value");
    assert_equals(w.opener, null, "'opener' property value");
    assert_equals(w.length, 1, "'length' property value");
    assert_not_throwing(function() { w[0]; }, "Subframe access via index");
    assert_equals(w['subframe'], w[0], "Subframe access by name");
}

function checkPopupPropertyValues(w)
{
    assert_equals(w.parent, w, "'parent' property value");
    assert_equals(w.top, w, "'top' property value");
    assert_equals(w.opener, window, "'opener' property value");
    assert_equals(w.length, 1, "'length' property value");
    assert_not_throwing(function() { w[0]; }, "Subframe access via index");
    assert_equals(w['subframe'], w[0], "Subframe access by name");
}

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=deny", true /* isCrossOrigin */).then((f) => {
        testCrossOriginOption(f.contentWindow, "deny", true /* isCrossOrigin */);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: deny' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=allow-postmessage", true /* isCrossOrigin */).then((f) => {
        testCrossOriginOption(f.contentWindow, "allow-postmessage", true /* isCrossOrigin */);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: allow-postmessage' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=alLoW-postMessAgE", true /* isCrossOrigin */).then((f) => {
        testCrossOriginOption(f.contentWindow, "allow-postmessage", true /* isCrossOrigin */);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: alLoW-postMessAgE' HTTP header (mixed case)");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=deny,allow", true /* isCrossOrigin */).then((f) => {
        const w = f.contentWindow;
        // Invalid input: should be treated as "allow".
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: deny,allow' HTTP header (multiple values is invalid)");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=", true /* isCrossOrigin */).then((f) => {
        const w = f.contentWindow;
        // Empty value: should be treated as "allow".
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options:' HTTP header (empty value)");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=allow", true /* isCrossOrigin */).then((f) => {
        const w = f.contentWindow;
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: allow' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=invalid", true /* isCrossOrigin */).then((f) => {
        const w = f.contentWindow;
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);
        
        checkIframePropertyValues(w);
    });
}, "Cross-origin iframe with 'Cross-Origin-Options: invalid' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=deny", false /* isCrossOrigin */).then((f) => {
         const w = f.contentWindow;
        testCrossOriginOption(w, "deny", false /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Same-origin iframe with 'Cross-Origin-Options: deny' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=allow-postmessage", false /* isCrossOrigin */).then((f) => {
         const w = f.contentWindow;
        testCrossOriginOption(w, "allow-postmessage", false /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Same-origin iframe with 'Cross-Origin-Options: allow-postmessage' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=allow", false /* isCrossOrigin */).then((f) => {
         const w = f.contentWindow;
        testCrossOriginOption(w, "allow", false /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Same-origin iframe with 'Cross-Origin-Options: allow' HTTP header");

promise_test(function(test) {
    return withIframe("serve-cross-origin-options-header.py?value=invalid", false /* isCrossOrigin */).then((f) => {
         const w = f.contentWindow;
        testCrossOriginOption(w, "allow", false /* isCrossOrigin */);

        checkIframePropertyValues(w);
    });
}, "Same-origin iframe with 'Cross-Origin-Options: invalid' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=deny", true /* isCrossOrigin */).then((result) => {
        testCrossOriginOption(result.window, "deny", true /* isCrossOrigin */);
    });
}, "Cross-origin popup with 'Cross-Origin-Options: deny' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=allow-postmessage", true /* isCrossOrigin */).then((result) => {
        testCrossOriginOption(result.window, "allow-postmessage", true /* isCrossOrigin */);
    });
}, "Cross-origin popup with 'Cross-Origin-Options: allow-postmessage' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=allow", true /* isCrossOrigin */).then((result) => {
        const w = result.window;
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Cross-origin popup with 'Cross-Origin-Options: allow' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=invalid", true /* isCrossOrigin */).then((result) => {
        const w = result.window;
        testCrossOriginOption(w, "allow", true /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Cross-origin popup with 'Cross-Origin-Options: invalid' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=deny", false /* isCrossOrigin */).then((result) => {
         const w = result.window;
        testCrossOriginOption(w, "deny", false /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Same-origin popup with 'Cross-Origin-Options: deny' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=allow-postmessage", false /* isCrossOrigin */).then((result) => {
         const w = result.window;
        testCrossOriginOption(w, "allow-postmessage", false /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Same-origin popup with 'Cross-Origin-Options: allow-postmessage' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=allow", false /* isCrossOrigin */).then((result) => {
         const w = result.window;
        testCrossOriginOption(w, "allow", false /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Same-origin popup with 'Cross-Origin-Options: allow' HTTP header");

promise_test(function(test) {
    return withPopup("serve-cross-origin-options-header.py?value=invalid", false /* isCrossOrigin */).then((result) => {
         const w = result.window;
        testCrossOriginOption(w, "allow", false /* isCrossOrigin */);

        checkPopupPropertyValues(w);
    });
}, "Same-origin popup with 'Cross-Origin-Options: invalid' HTTP header");

</script>
</body>
</html>
